+++ src/main/java/lk/ijse/green_shadow_backend/controller/FieldController.java
@@ -3,6 +3,7 @@ package lk.ijse.green_shadow_backend.controller;
 import lk.ijse.green_shadow_backend.dto.impl.FieldDTO;
 import lk.ijse.green_shadow_backend.service.FieldService;
 import lk.ijse.green_shadow_backend.util.AppUtil;
+import lk.ijse.green_shadow_backend.util.Regex;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -39,8 +40,12 @@ public class FieldController {
     @PreAuthorize("hasRole('MANAGER') or hasRole('SCIENTIST')")
     public ResponseEntity<Void> updateField(@PathVariable("fieldCode") String fieldCode,@RequestBody FieldDTO fieldDTO){
         try{
-            fieldService.updateField(fieldCode,fieldDTO);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if(Regex.FIELD_CODE.validate(fieldCode)){
+                fieldService.updateField(fieldCode,fieldDTO);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }else {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
         }catch (Exception e){
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -50,8 +55,12 @@ public class FieldController {
     @PreAuthorize("hasRole('MANAGER') or hasRole('SCIENTIST')")
     public ResponseEntity<Void> deleteField(@PathVariable("fieldCode") String fieldCode){
         try{
-            fieldService.deleteField(fieldCode);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if(Regex.FIELD_CODE.validate(fieldCode)){
+                fieldService.deleteField(fieldCode);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }else{
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
         }catch (Exception e){
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -65,10 +74,14 @@ public class FieldController {
             @RequestPart("image2") MultipartFile image2
     ){
         try{
-            byte[] image1Bytes = image1.getBytes();
-            byte[] image2Bytes = image2.getBytes();
-            fieldService.uploadFieldImage(fieldCode, AppUtil.convertImageToBase64(image1Bytes), AppUtil.convertImageToBase64(image2Bytes));
-            return new ResponseEntity<>(HttpStatus.CREATED);
+            if(Regex.FIELD_CODE.validate(fieldCode)){
+                byte[] image1Bytes = image1.getBytes();
+                byte[] image2Bytes = image2.getBytes();
+                fieldService.uploadFieldImage(fieldCode, AppUtil.convertImageToBase64(image1Bytes), AppUtil.convertImageToBase64(image2Bytes));
+                return new ResponseEntity<>(HttpStatus.CREATED);
+            }else{
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
         }catch (Exception e){
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
